chore(cgroup): add condition compile for IsCgroupPath
